move EngineFilter classes to their own folder
