+++ src/effects/builtin/balanceeffect.h
@@ -4,7 +4,7 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterlinkwitzriley4.h"
+#include "engine/filters/enginefilterlinkwitzriley4.h"
 #include "util/samplebuffer.h"
 #include "util/memory.h"
