+++ src/effects/builtin/bessel4lvmixeqeffect.h
@@ -9,8 +9,8 @@
 #include "effects/builtin/lvmixeqbase.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterbessel4.h"
-#include "engine/enginefilterdelay.h"
+#include "engine/filters/enginefilterbessel4.h"
+#include "engine/filters/enginefilterdelay.h"
 #include "util/class.h"
 #include "util/types.h"
 #include "util/defs.h"
