+++ src/effects/builtin/bessel8lvmixeqeffect.h
@@ -10,8 +10,8 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterbessel8.h"
-#include "engine/enginefilterdelay.h"
+#include "engine/filters/enginefilterbessel8.h"
+#include "engine/filters/enginefilterdelay.h"
 #include "util/class.h"
 #include "util/defs.h"
 #include "util/sample.h"
