+++ src/effects/builtin/biquadfullkilleqeffect.h
@@ -6,10 +6,10 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterbiquad1.h"
-#include "engine/enginefilterbessel4.h"
+#include "engine/filters/enginefilterbiquad1.h"
+#include "engine/filters/enginefilterbessel4.h"
 #include "effects/builtin/lvmixeqbase.h"
-#include "engine/enginefilterdelay.h"
+#include "engine/filters/enginefilterdelay.h"
 #include "util/class.h"
 #include "util/defs.h"
 #include "util/sample.h"
