+++ src/effects/builtin/linkwitzriley8eqeffect.h
@@ -8,7 +8,7 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterlinkwitzriley8.h"
+#include "engine/filters/enginefilterlinkwitzriley8.h"
 #include "util/class.h"
 #include "util/defs.h"
 #include "util/sample.h"
