+++ src/effects/builtin/metronomeeffect.h
@@ -6,7 +6,7 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefilterpansingle.h"
+#include "engine/filters/enginefilterpansingle.h"
 #include "util/class.h"
 #include "util/defs.h"
 #include "util/sample.h"
