+++ src/effects/builtin/moogladder4filtereffect.h
@@ -5,7 +5,7 @@
 #include "effects/effectprocessor.h"
 #include "engine/effects/engineeffect.h"
 #include "engine/effects/engineeffectparameter.h"
-#include "engine/enginefiltermoogladder4.h"
+#include "engine/filters/enginefiltermoogladder4.h"
 #include "util/class.h"
 #include "util/defs.h"
 #include "util/sample.h"
