+++ src/engine/filters/enginefilterbessel4.h
@@ -1,7 +1,7 @@
 #ifndef ENGINEFILTERBESSEL4_H
 #define ENGINEFILTERBESSEL4_H
 
-#include "engine/enginefilteriir.h"
+#include "engine/filters/enginefilteriir.h"
 
 class EngineFilterBessel4Low : public EngineFilterIIR<4, IIR_LP> {
     Q_OBJECT
