+++ src/engine/filters/enginefilterbessel8.h
@@ -1,7 +1,7 @@
 #ifndef ENGINEFILTERBESSEL8_H
 #define ENGINEFILTERBESSEL8_H
 
-#include "engine/enginefilteriir.h"
+#include "engine/filters/enginefilteriir.h"
 
 class EngineFilterBessel8Low : public EngineFilterIIR<8, IIR_LP> {
     Q_OBJECT
