+++ src/engine/filters/enginefilterbiquad1.h
@@ -1,7 +1,7 @@
 #ifndef ENGINEFILTERBIQUAD1_H
 #define ENGINEFILTERBIQUAD1_H
 
-#include "engine/enginefilteriir.h"
+#include "engine/filters/enginefilteriir.h"
 
 #ifdef _MSC_VER
     // Visual Studio doesn't have snprintf
