+++ src/engine/filters/enginefilterpansingle.h
