+++ src/preferences/dialog/dlgprefeq.cpp
@@ -23,7 +23,7 @@
 #include "effects/builtin/biquadfullkilleqeffect.h"
 #include "effects/builtin/filtereffect.h"
 #include "effects/effectslot.h"
-#include "engine/enginefilterbessel4.h"
+#include "engine/filters/enginefilterbessel4.h"
 #include "control/controlobject.h"
 #include "control/controlproxy.h"
 #include "util/math.h"
