+++ src/test/enginefilterbiquadtest.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "engine/enginefilterbiquad1.h"
+#include "engine/filters/enginefilterbiquad1.h"
 
 namespace {
